Remove optional show comments, adjust read me, version bump
This did not work. Reason is that if an "outer" cache was being written while the "inner" didn't show its comments, the outer of course wouldn't know about it.

Until I come up with a working way to conditionally show comments, just show them all the time. We all have broadband anyway.